Add all CRUD functionalities concerning LOG

+++ management_project/management_project/Structures.h
@@ -63,7 +63,7 @@ struct LOG
 {
 	int id = 0;
 	int userId = 0;
-	int timeSpent = 0;
+	double timeSpent = 0;
 	std::string date = "";
 	int taskId = 0;
